add removeProduct

+++ components/card.tsx
@@ -35,11 +35,15 @@ const Card: NextPage = () => {
   }, []);
 
   const handleRemoveProduct = () => {
+    
     try {
-      axios.delete(`http://localhost:3333/products/${product.id}.json`)
-      .then((res) => {
-        return res.data;
-      })   
+      if(window.confirm('Are you sure you want to delete')){
+
+        axios.delete(`http://localhost:3333/products/${product.id}.json`)
+        .then((res) => {
+          return res.data;
+        })   
+      }
   }catch(error) {
     console.log('Error to delete');
   }
